103 debug early batch of seeder
