+++ packages/backend/src/seeds/DatabaseSeeds.ts
@@ -1,4 +1,4 @@
-import { ISeeds } from '@Interfaces/ISeeds';
+import { ISeeds } from '../libs/interfaces/ISeeds';
 import { getConnection } from 'typeorm';
 import { OrganisationSeeds } from './OrganisationSeeds';
 import { StaffSeeds } from './StaffSeeds';
@@ -7,7 +7,10 @@ import { UserSeeds } from './UserSeeds';
 export class DatabaseSeeds implements ISeeds
 {
     async run(): Promise<void> {
+        console.log('-----------Cleaning Database');
         this.clearDatabase();
+
+        console.log('-----------Seeding database');
         await (new OrganisationSeeds()).run();
         await (new UserSeeds()).run();
         await (new StaffSeeds()).run();
@@ -19,7 +22,7 @@ export class DatabaseSeeds implements ISeeds
 
         entities.forEach(async (entity) => {
             const repository = connection.getRepository(entity.name);
-            await repository.query(`DELETE FROM ${entity.tableName}`);
+            await repository.query(`DELETE FROM public.${entity.tableName}`);
         });
     }
