+++ packages/backend/tsconfig.json
@@ -13,11 +13,6 @@
     "baseUrl": ".",
     "paths": {
       "*": ["node_modules/*", "src/types/*"],
-      "@ClugBackend/*": ["./src/*"], 
-      "@Models/*": ["./src/Models/*"],
-      "@Classes/*": ["./src/libs/classes/*"],
-      "@Interfaces/*": ["./src/libs/interfaces/*"],
-      "@Controllers/*": ["./src/controllers/*"],
     }
   },
   "include": ["src/**/*"],
